video_production_skills/depth_parallax/requirements.txt
torch>=2.1
torchvision>=0.16
timm>=0.9
opencv-python>=4.8
Pillow>=10.0
numpy>=1.24
